Replace manual action scale with auto
Automatically calculate the action offset and scale
from joint limits. Compared to previous version, the
scale is lower for the hip and calf DOFs but higher for
the knee DOF.

Note: Computing observations still uses default_dof_pos
and default_dof_vel. This shouldn't really be a problem
since it's just a linear transformation in each dimension
but it's something to fix in the future.